Added a strategy for procurement of resources

+++ Spaceship.h
@@ -51,7 +51,7 @@ class Spaceship {
 
 
     private:
-        int resources;/**<Number of resources on the ship*/
+        double resources;/**<Number of resources on the ship*/
         int displacement;/**<displacement of spaceship*/
         int power;/**<power of spaceship */
         int thrust;/**<thrust*/
